move dialog component
